use hashes in the configmap names (#47)
When updating a ConfigMap, there is no mechanism in place that will trigger the Pods which rely on that ConfigMap to redeploy. This PR adds a random character string into the ConfigMap name so that when the template is reapplied, it will generate a new character string changing the name of the ConfigMap. Since trino will now reference a new ConfigMap, the pods will redeploy picking up the new changes.